Attempt 1 at fixing black models with Blend shader
I made a quick change that may or may not have been part of the issue.

Let me know if it fixes it by chance.

And if not, then one more thing to try would be to remove the base normal map from the Erika model, and then left me know if that gets the black-rendering to go away. 

Also, if the issue still persists after this PR, it could be useful to set the "DebugValuesMode" MatParam and take a screenshot of what it looks like showing each individual layer.

That way I can know if it is related to the final lighting calculation, or if it is an issue with a single broken ao/normal layer, since bad normals or an AO value of 0.0 will often cause fully black rendering. But it could also be any number of variables that get sent into the final lighting equation that are not having the proper value set.